Rename classes * Booking -> BookingEntity * Package -> PackageEntity * Pass -> PassEntity * User -> UserEntity
