--- src/main/java/com/dev/passbatch/repository/BookingRepository.java
@@ -1,7 +1,0 @@
-package com.dev.passbatch.repository;
-
-import com.dev.passbatch.domain.Booking;
-import org.springframework.data.jpa.repository.JpaRepository;
-
-public interface BookingRepository extends JpaRepository<Booking, Long> {
-}
