--- src/main/java/com/dev/passbatch/repository/PackageRepository.java
@@ -1,7 +1,0 @@
-package com.dev.passbatch.repository;
-
-import com.dev.passbatch.domain.Package;
-import org.springframework.data.jpa.repository.JpaRepository;
-
-public interface PackageRepository extends JpaRepository<Package, Long> {
-}
