--- src/main/java/com/dev/passbatch/repository/PassRepository.java
@@ -1,7 +1,0 @@
-package com.dev.passbatch.repository;
-
-import com.dev.passbatch.domain.Pass;
-import org.springframework.data.jpa.repository.JpaRepository;
-
-public interface PassRepository extends JpaRepository<Pass, Long> {
-}
