--- src/main/java/com/dev/passbatch/repository/UserRepository.java
@@ -1,7 +1,0 @@
-package com.dev.passbatch.repository;
-
-import com.dev.passbatch.domain.User;
-import org.springframework.data.jpa.repository.JpaRepository;
-
-public interface UserRepository extends JpaRepository<User, Long> {
-}
