+++ src/main/java/com/dev/passbatch/repository/booking/BookingRepository.java
@@ -1,0 +1,6 @@
+package com.dev.passbatch.repository.booking;
+
+import org.springframework.data.jpa.repository.JpaRepository;
+
+public interface BookingRepository extends JpaRepository<BookingEntity, Integer> {
+}
