+++ src/main/java/com/dev/passbatch/repository/packaze/PackageRepository.java
@@ -1,0 +1,6 @@
+package com.dev.passbatch.repository.packaze;
+
+import org.springframework.data.jpa.repository.JpaRepository;
+
+public interface PackageRepository extends JpaRepository<PackageEntity, Integer> {
+}
