+++ src/main/java/com/dev/passbatch/repository/pass/PassRepository.java
@@ -1,0 +1,6 @@
+package com.dev.passbatch.repository.pass;
+
+import org.springframework.data.jpa.repository.JpaRepository;
+
+public interface PassRepository extends JpaRepository<PassEntity, Integer> {
+}
