+++ src/main/java/com/dev/passbatch/repository/user/UserRepository.java
@@ -1,0 +1,6 @@
+package com.dev.passbatch.repository.user;
+
+import org.springframework.data.jpa.repository.JpaRepository;
+
+public interface UserRepository extends JpaRepository<UserEntity, Integer> {
+}
